requirements.txt
yfinance
gspread
oauth2client
python-dotenv
pandas
pytest
